Extract PivotRunsTable
